css-text/line-breaking/css-text-line-break-ba-blocks.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>Line-breaking: By block</title>
<link rel="author" title="Richard Ishida" href="mailto:ishida@w3.org">
<link rel="help" href="https://drafts.csswg.org/css-text-3/#line-breaking">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<meta name="assert" content="[Exploratory] The browser will break a line of text after each of the BA characters in the script blocks listed.">
<style type="text/css">
.breakable { font-size: 25px; width: 800px; line-height: 30px; }
</style>
<script>
var charlist = `1173C AHOM SIGN SMALL SECTION
1173D AHOM SIGN SECTION
1173E AHOM SIGN RULAI
058A ARMENIAN HYPHEN
1B5E BALINESE CARIK SIKI
1B5F BALINESE CARIK PAREREN
1B5A BALINESE PANTI
1B5B BALINESE PAMADA
1B5D BALINESE CARIK PAMUNGKAH
1B60 BALINESE PAMENENG
A6F3 BAMUM FULL STOP
A6F4 BAMUM COLON
A6F5 BAMUM COMMA
A6F6 BAMUM SEMICOLON
A6F7 BAMUM QUESTION MARK
16AF5 BASSA VAH FULL STOP
11C41 BHAIKSUKI DANDA
11C42 BHAIKSUKI DOUBLE DANDA
11C43 BHAIKSUKI WORD SEPARATOR
11C44 BHAIKSUKI GAP FILLER-1
11C45 BHAIKSUKI GAP FILLER-2
1400 CANADIAN SYLLABICS HYPHEN
11140 CHAKMA SECTION MARK
11141 CHAKMA DANDA
11142 CHAKMA DOUBLE DANDA
11143 CHAKMA QUESTION MARK
AA5D CHAM PUNCTUATION DANDA
AA5E CHAM PUNCTUATION DOUBLE DANDA
AA5F CHAM PUNCTUATION TRIPLE DANDA
2CFA COPTIC OLD NUBIAN DIRECT QUESTION MARK
2CFB COPTIC OLD NUBIAN INDIRECT QUESTION MARK
2CFC COPTIC OLD NUBIAN VERSE DIVIDER
2CFF COPTIC MORPHOLOGICAL DIVIDER
0964 DEVANAGARI DANDA
0965 DEVANAGARI DOUBLE DANDA
05BE HEBREW PUNCTUATION MAQAF
1361 ETHIOPIC WORDSPACE
A9C7 JAVANESE PADA PANGKAT
A9C8 JAVANESE PADA LINGSA
A9C9 JAVANESE PADA LUNGSI
110BE KAITHI SECTION MARK
110BF KAITHI DOUBLE SECTION MARK
110C0 KAITHI DANDA
110C1 KAITHI DOUBLE DANDA
A92E KAYAH LI SIGN CWI
A92F KAYAH LI SIGN SHYA
17D8 KHMER SIGN BEYYAL
17DA KHMER SIGN KOOMUUT
17D4 KHMER SIGN KHAN
17D5 KHMER SIGN BARIYOOSAN
11238 KHOJKI DANDA
11239 KHOJKI DOUBLE DANDA
1123B KHOJKI SECTION MARK
1123C KHOJKI DOUBLE SECTION MARK
1C3B LEPCHA PUNCTUATION TA-ROL
1C3C LEPCHA PUNCTUATION NYET THYOOM TA-ROL
1C3D LEPCHA PUNCTUATION CER-WA
1C3E LEPCHA PUNCTUATION TSHOOK CER-WA
1C3F LEPCHA PUNCTUATION TSHOOK
A4FE LISU PUNCTUATION COMMA
A4FF LISU PUNCTUATION FULL STOP
AAF0 MEETEI MAYEK CHEIKHAN
AAF1 MEETEI MAYEK AHANG KHUDAM
ABEB MEETEI MAYEK CHEIKHEI
11641 MODI DANDA
11642 MODI DOUBLE DANDA
1804 MONGOLIAN COLON
1805 MONGOLIAN FOUR DOTS
16A6E MRO DANDA
16A6F MRO DOUBLE DANDA
112A9 MULTANI SECTION MARK
104A MYANMAR SIGN LITTLE SECTION
104B MYANMAR SIGN SECTION
1144B NEWA DANDA
1144C NEWA DOUBLE DANDA
1144D NEWA COMMA
1144E NEWA GAP FILLER
1145B NEWA PLACEHOLDER MARK
1C7E OL CHIKI PUNCTUATION MUCAAD
1C7F OL CHIKI PUNCTUATION DOUBLE MUCAAD
16B37 PAHAWH HMONG SIGN VOS THOM
16B38 PAHAWH HMONG SIGN VOS TSHAB CEEB
16B39 PAHAWH HMONG SIGN CIM CHEEM
16B44 PAHAWH HMONG SIGN XAUS
1735 PHILIPPINE SINGLE PUNCTUATION
1736 PHILIPPINE DOUBLE PUNCTUATION
16EB RUNIC SINGLE PUNCTUATION
16EC RUNIC MULTIPLE PUNCTUATION
16ED RUNIC CROSS PUNCTUATION
A8CE SAURASHTRA DANDA
A8CF SAURASHTRA DOUBLE DANDA
111C5 SHARADA DANDA
111C6 SHARADA DOUBLE DANDA
111C8 SHARADA SEPARATOR
111DD SHARADA CONTINUATION SIGN
111DE SHARADA SECTION MARK-1
111DF SHARADA SECTION MARK-2
115C2 SIDDHAM DANDA
115C3 SIDDHAM DOUBLE DANDA
115C9 SIDDHAM END OF TEXT MARK
115CA SIDDHAM SECTION MARK WITH TRIDENT AND U-SHAPED ORNAMENTS
115CB SIDDHAM SECTION MARK WITH TRIDENT AND DOTTED CRESCENTS
115CC SIDDHAM SECTION MARK WITH RAYS AND DOTTED CRESCENTS
115CD SIDDHAM SECTION MARK WITH RAYS AND DOTTED DOUBLE CRESCENTS
115CE SIDDHAM SECTION MARK WITH RAYS AND DOTTED TRIPLE CRESCENTS
115CF SIDDHAM SECTION MARK DOUBLE RING
115D0 SIDDHAM SECTION MARK DOUBLE RING WITH RAYS
115D1 SIDDHAM SECTION MARK WITH DOUBLE CRESCENTS
115D2 SIDDHAM SECTION MARK WITH TRIPLE CRESCENTS
115D3 SIDDHAM SECTION MARK WITH QUADRUPLE CRESCENTS
115D4 SIDDHAM SECTION MARK WITH SEPTUPLE CRESCENTS
115D5 SIDDHAM SECTION MARK WITH CIRCLES AND RAYS
115D6 SIDDHAM SECTION MARK WITH CIRCLES AND TWO ENCLOSURES
115D7 SIDDHAM SECTION MARK WITH CIRCLES AND FOUR ENCLOSURES
0F0B TIBETAN MARK INTERSYLLABIC TSHEG
0F34 TIBETAN MARK BSDUS RTAGS
0F7F TIBETAN SIGN RNAM BCAD
0F85 TIBETAN MARK PALUTA
0FBE TIBETAN KU RU KHA
0FBF TIBETAN KU RU KHA BZHI MIG CAN
0FD2 TIBETAN MARK NYIS TSHEG
2D70 TIFINAGH SEPARATOR MARK
0E5A THAI CHARACTER ANGKHANKHU
0E5B THAI CHARACTER KHOMUT
A60D VAI COMMA
A60F VAI QUESTION MARK`
</script>
</head>
<body>
<script>
var lines = charlist.split('\n')
var out = '<div id="log"></div>\n'
for (var i=0;i<lines.length;i++) {
	// get the data
	var firstSpace = lines[i].indexOf(' ')
	var hex = lines[i].substr(0,firstSpace)
	var name = lines[i].substr(firstSpace)
	// make a test
	out +=  '<div><div>'+hex+'</div><div id="test'+i+'" class="breakable">aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa&#x'+hex+';bbbbbbbbbbbbbbbbbbbbbbbbbbbbbb</div></div>'
	}
document.querySelector('body').innerHTML = out
// run the test framework
 for (i=0;i<lines.length;i++) {
   test(function() {
        assert_true(document.getElementById('test'+i).offsetHeight > 35);
        }, lines[i]+' should allow a line break after');
   }
// hide successful tests
for (i=0;i<lines.length;i++) {
    if (document.getElementById('test'+i).offsetHeight > 35) document.getElementById('test'+i).parentNode.style.display = 'none'
    }
</script>
<!--Notes:
This list is based on property assignments for Unicode version 10.
Each test creates an unbroken text item from two very long words connected by the character being tested. The whole text item is displayed in a box that is narrower than the width of the text item.  If the browser allows a line to wrap after the character being tested, the text item will break.  The test measures the height of the text box to determine whether the text wrapped.
-->
</body>
</html>
